Fixed bug

+++ RegKeys/Form1.cs
@@ -24,40 +24,55 @@ namespace RegKeys
             Form2 f2 = new Form2();
             f2.ShowDialog();
             namereg = f2.NameOfReg;
-            for (int i = 0; i < RegistryList.Items.Count; i++)
+            if (namereg != null)
             {
-                if (namereg == RegistryList.Items[i].ToString()) { RegistryList.Items.RemoveAt(i); break; }
+                for (int i = 0; i < RegistryList.Items.Count; i++)
+                {
+                    if (namereg == RegistryList.Items[i].ToString()) { RegistryList.Items.RemoveAt(i); break; }
+                }
+                RegistryList.Items.Add(namereg);
             }
-            RegistryList.Items.Add(namereg);
         }
 
         private void GetRegVal_Click(object sender, EventArgs e)
         {
             RegTree.Items.Clear();
-            RegistryKey key = Registry.CurrentUser.OpenSubKey(Path.Combine("Software", RegistryList.SelectedItem.ToString()));
-            Dictionary<string, object> keyValuePairs;
-            RegTree.Items.Add(key.View.ToString());
-            using (var settingsRegKey = Registry.CurrentUser.OpenSubKey(Path.Combine("Software", namereg)))
+            if (RegistryList.SelectedItem != null)
             {
-                var valueNames = settingsRegKey.GetValueNames();
-                keyValuePairs = valueNames.ToDictionary(name => name, settingsRegKey.GetValue);
-                foreach (var regkey in keyValuePairs)
+                RegistryKey key = Registry.CurrentUser.OpenSubKey(Path.Combine("Software", RegistryList.SelectedItem.ToString()));
+                Dictionary<string, object> keyValuePairs;
+                RegTree.Items.Add(key.View.ToString());
+                using (var settingsRegKey = Registry.CurrentUser.OpenSubKey(Path.Combine("Software", namereg)))
                 {
-                    RegTree.Items.Add(regkey.ToString() + settingsRegKey.GetValue(regkey.ToString()));
+                    var valueNames = settingsRegKey.GetValueNames();
+                    keyValuePairs = valueNames.ToDictionary(name => name, settingsRegKey.GetValue);
+                    foreach (var regkey in keyValuePairs)
+                    {
+                        RegTree.Items.Add(regkey.ToString() + settingsRegKey.GetValue(regkey.ToString()));
+                    }
                 }
             }
+            
         }
 
         private void DeleteReg_Click(object sender, EventArgs e)
         {
-            Registry.CurrentUser.DeleteSubKeyTree(Path.Combine("Software", RegistryList.SelectedItem.ToString()));
-            RegistryList.Items.RemoveAt(RegistryList.SelectedIndex);
-            RegTree.Items.Clear();
+            if (RegistryList.SelectedItem != null)
+            {
+                Registry.CurrentUser.DeleteSubKeyTree(Path.Combine("Software", RegistryList.SelectedItem.ToString()));
+                RegistryList.Items.RemoveAt(RegistryList.SelectedIndex);
+                RegTree.Items.Clear();
+            }
         }
 
         private void Clear_RegTree_Click(object sender, EventArgs e)
         {
             RegTree.Items.Clear();
+        }
+
+        private void Form1_Load(object sender, EventArgs e)
+        {
+
         }
     }
 }
